The initial version of imap library for Gmail.
The is not a general imap library, but a special imap
client implementation for Gmail imap server. In other words,
this code is supposed to work with Gmail imap server(s) only.

GmailLib implements the basic imap commands to read information from
Gmail account. Important note is that imap support should be enabled
in account, otherwise it is impossible to establish imap connection
to the server.