Refactor server response handling code.
The previous implementation worked incorrectly and had needless
buffer storage in the command structure. It turned out that in
some cases server can send responses from the different commands
in the same chunk of data and we did not correctly parse it. As a
result commands handling architecture flawed with inconsistent
content of command/response mapping.
The new code will/should handle such cases in correct way.
We also get rid of keeping buffer in the command structure.