Get rid of using QEventLoop when wait for user is logged in.
Using QEventLoop leads to inconsistent results when quiting event
loop was not really stop it, so executing it in that very moment
block the event loop. This caused gherald to stop processing commands
correctly.
With this change we replaced event loop with while loop with timeout.
Also improved logged in state handling.